+++ mmodel/metadata.py
@@ -54,25 +54,60 @@ def format_func(key, value):
     return [f"{value.__name__}{inspect.signature(value)}"]
 
 
-def format_listargs(key, value):
-    """Format the metadata value that has a list of (func, kwargs).
-
-    The formatter is for modifiers and other metadata that have a
-    list of (func, kwargs), and kwargs is a dictionary.
-    """
+def format_list(key, value):
+    """Format the metadata value that is a list."""
 
     if value:
         str_list = [f"{key}:"]
-
-        for func, kwargs in value:
-            str_value = [repr(v) for v in kwargs.values()]
-            mod_str = f"\t- {func.__name__}({', '.join(str_value)})"
-            str_list.append(mod_str)
-
+        for v in value:
+            str_list.append(f"\t- {v}")
         return str_list
-
     else:
         return []
+
+
+def modifier_metadata(closure):
+    """Extract metadata from closure, including the name and the arguments.
+
+    The order of extraction:
+    1. If the object has the "metadata" attribute defined.
+    2. If the closure takes no arguments, the name is the function name.
+    3. If the closure takes arguments, the "metadata" attribute is not defined.
+
+    Note::
+
+        inspect.getclosurevars(closure).nonlocals can only parse values
+        if the value is used in the closure.
+    """
+
+    if hasattr(closure, "metadata"):
+        return closure.metadata
+    elif not inspect.getclosurevars(closure).nonlocals:
+        return closure.__name__
+
+    else:  # closure takes arguments
+
+        # In some rare cases, the closure is a nested function.
+        # For example, in the tests, the nested closure reflects the
+        # path of the parent function. Here we remove the nested
+        # parent function name.
+
+        name = closure.__qualname__.rsplit(".<locals>.")[-2]
+        kwargs = inspect.getclosurevars(closure).nonlocals
+        kwargs_str = ", ".join(f"{k}={repr(v)}" for k, v in kwargs.items())
+        return f"{name}({kwargs_str})"
+
+
+def format_modifierlist(key, value):
+    """Format the metadata that is a list of modifiers.
+
+    The metadata of the modifier is extracted by the modifier_metadata function.
+    The resulting list is formatted by the format_list function.
+    """
+
+    modifier_str_list = [modifier_metadata(modifier) for modifier in value]
+
+    return format_list(key, modifier_str_list)
 
 
 def format_dictargs(key, value):
@@ -154,7 +189,7 @@ modelformatter = MetaDataFormatter(
         "graph": format_obj,
         "handler": format_obj,
         "handler args": format_dictargs,
-        "modifiers": format_listargs,
+        "modifiers": format_modifierlist,
         "description": format_value,
     },
     [
@@ -173,7 +208,7 @@ nodeformatter = MetaDataFormatter(
     {
         "node": format_value,
         "func": format_func,
-        "modifiers": format_listargs,
+        "modifiers": format_modifierlist,
         "doc": format_value,
     },
     ["node", None, "func", "return", "functype", "modifiers", None, "doc"],